Created a strategies package.
